fixed 64 bit compile errors and gcc4 errors

+++ src/cmd/missile_generic.cpp
@@ -134,7 +134,7 @@ void Missile::UpdatePhysics2 (const Transformation &trans, const Transformation 
 			targ=NULL; 
 		}else {
                   static int max_ecm = XMLSupport::parse_int(vs_config->getVariable("physics","max_ecm","4"));
-			if ((((int)this)/16383)%max_ecm<targ->GetImageInformation().ecm){
+			if ((((size_t)this)/16383)%max_ecm<targ->GetImageInformation().ecm){
 				Target (NULL);//go wild
 			}else if (hull>0){
 				static unsigned int pointdef = ROLES::getRole("POINTDEF");
